+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- [JAF-231](https://makeitapp.atlassian.net/browse/JAF-231): added support for `find_one` builtin in Rego policies
+
 ## 0.3.1 - 31/01/2022
 
 ### Changed